requirements/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mock==5.1.0
mysqlclient==2.2.3
packaging==23.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
structlog==24.1.0
